More rearranging, don't rely on system locale
Rearrange code between modules and replace use of Data.Text.IO with
IO from Data.ByteString followed by encode/decode, to ensure that utf-8
is always used.
