+++ src/PscPages/AsHtml.hs
@@ -1,19 +1,30 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+-- | Functions for rendering generated documentation from PureScript code as
+-- HTML.
 
 module PscPages.AsHtml where
 
--- | Functions for rendering generated documentation from PureScript code as
--- | HTML.
-
+import Control.Applicative
+import Data.Ord (comparing)
+import Data.Char (toUpper)
 import Data.String (fromString)
 import Data.Foldable (for_)
 import Data.Maybe (fromMaybe)
-import Data.List (intercalate)
+import Data.List (nub, intercalate, sortBy)
 import Data.List.Split (splitOn)
+
+import qualified Data.ByteString as B
+import qualified Data.Text as T
+import qualified Data.Text.Lazy as TL
+import qualified Data.Text.Encoding as TE
+import qualified Data.Text.Lazy.Encoding as TLE
 
 import Text.Blaze.Html ((!))
 import qualified Text.Blaze.Html as H
+import qualified Text.Blaze.Html.Renderer.Text as H
 import qualified Text.Blaze.Html5 as H
 import qualified Text.Blaze.Html5.Attributes as A
 
@@ -21,15 +32,79 @@ import qualified Language.PureScript as P
 
 import System.FilePath ((</>))
 
+import Data.FileEmbed
+
 import PscPages.HtmlHelpers
 import PscPages.RenderedCode hiding (sp)
 import PscPages.Render
+import PscPages.Output
+import PscPages.IOUtils
 
--- A tuple containing:
---    - bookmarks,
---    - source module name (that is, the name of the module which is currently
---      being generated)
-type LinksContext = ([(P.ModuleName, String)], P.ModuleName)
+-- | A tuple containing:
+-- * bookmarks,
+-- * source module name (that is, the name of the module which is currently
+--   being generated).
+type LinksContext = (Bookmarks, P.ModuleName)
+
+outputAsHtml :: FilePath -> OutputFn
+outputAsHtml outputDir bookmarks modules = do
+  let stylesheetFile = outputDir </> "style.css"
+  mkdirp stylesheetFile
+  writeTextFile stylesheetFile stylesheet
+
+  let bootstrapFile = outputDir </> "bootstrap.min.css"
+  mkdirp bootstrapFile
+  writeTextFile bootstrapFile bootstrap
+
+  let contentsFile = outputDir </> "index.html"
+  writeLazyTextFile contentsFile (H.renderHtml $ contentsPageHtml modules)
+
+  let indexFile = outputDir </> "index/index.html"
+  mkdirp indexFile
+  writeLazyTextFile indexFile (H.renderHtml indexPageHtml)
+
+  for_ ['a'..'z'] $ \c -> do
+    let letterFile = outputDir </> ("index/" ++ c : ".html")
+    writeLazyTextFile letterFile (H.renderHtml $ letterPageHtml c bookmarks)
+
+  for_ modules (renderModule' outputDir bookmarks)
+
+stylesheet :: T.Text
+stylesheet = TE.decodeUtf8 $(embedFile "static/style.css")
+
+bootstrap :: T.Text
+bootstrap = TE.decodeUtf8 $(embedFile "static/bootstrap.min.css")
+
+renderModule' :: FilePath -> [(P.ModuleName, String)] -> P.Module -> IO ()
+renderModule' outputDir bookmarks m@(P.Module _ moduleName _ _) = do
+  let filename = outputDir </> filePathFor moduleName
+      html = H.renderHtml $ moduleToHtml bookmarks m
+  mkdirp filename
+  writeLazyTextFile filename html
+
+contentsPageHtml :: [P.Module] -> H.Html
+contentsPageHtml ms = do
+  template "index.html" "Contents" $ do
+    H.h2 $ text "Modules"
+    H.ul $ for_ (sortBy (comparing $ \(P.Module _ moduleName _ _) -> moduleName) ms) $ \(P.Module _ moduleName _ _) -> H.li $
+      H.a ! A.href (fromString (filePathFor moduleName `relativeTo` "index.html")) $ text (show moduleName)
+
+indexPageHtml :: H.Html
+indexPageHtml = do
+  template "index/index.html" "Index" $ do
+    H.ul $ for_ ['a'..'z'] $ \c ->
+      H.li $ H.a ! A.href (fromString (c : ".html")) $ text [toUpper c]
+
+letterPageHtml :: Char -> [(P.ModuleName, String)] -> H.Html
+letterPageHtml c bs = do
+  let filename = "index/" ++ c : ".html"
+  template filename [toUpper c] $ do
+    H.ul $ for_ (sortBy (comparing snd) . nub . filter matches $ bs) $ \(mn, s) -> H.li $ H.code $ do
+      H.a ! A.href (fromString ((filePathFor mn `relativeTo` filename) ++ "#" ++ s)) $ text s
+      sp *> text ("(" ++ show mn ++ ")")
+  where
+  matches (_, (c':_)) = toUpper c == toUpper c'
+  matches _ = False
 
 template :: FilePath -> String -> H.Html -> H.Html
 template curFile title body = do
